+++ src/main/java/pro/sky/java/course3/task2/repositories/StudentRepository.java
@@ -6,11 +6,10 @@ import pro.sky.java.course3.task2.model.Student;
 import java.util.List;
 
 @Repository
-public interface StudentRepository extends JpaRepository<Student, Integer>{
+public interface StudentRepository extends JpaRepository<Student, Long>{
 
     List<Student> findByAge(Integer age);
 
     List<Student> findByAgeBetween(Integer min, Integer max);
 
-//    List<Student> findAllByFaculty(Integer id);
 }
